Updated models to have ImageManagement namespace
