+++ app/controllers/target_images_controller.rb
@@ -2,7 +2,7 @@ class TargetImagesController < ApplicationController
   # GET /target_images
   # GET /target_images.json
   def index
-    @target_images = TargetImage.all
+    @target_images = ImageManagement::TargetImage.all
 
     respond_to do |format|
       format.html # index.html.erb
@@ -13,7 +13,7 @@ class TargetImagesController < ApplicationController
   # GET /target_images/1
   # GET /target_images/1.json
   def show
-    @target_image = TargetImage.find(params[:id])
+    @target_image = ImageManagement::TargetImage.find(params[:id])
 
     respond_to do |format|
       format.html # show.html.erb
@@ -24,7 +24,7 @@ class TargetImagesController < ApplicationController
   # GET /target_images/new
   # GET /target_images/new.json
   def new
-    @target_image = TargetImage.new
+    @target_image = ImageManagement::TargetImage.new
 
     respond_to do |format|
       format.html # new.html.erb
@@ -34,13 +34,13 @@ class TargetImagesController < ApplicationController
 
   # GET /target_images/1/edit
   def edit
-    @target_image = TargetImage.find(params[:id])
+    @target_image = ImageManagement::TargetImage.find(params[:id])
   end
 
   # POST /target_images
   # POST /target_images.json
   def create
-    @target_image = TargetImage.new(params[:target_image])
+    @target_image = ImageManagement::TargetImage.new(params[:target_image])
 
     respond_to do |format|
       if @target_image.save
@@ -56,7 +56,7 @@ class TargetImagesController < ApplicationController
   # PUT /target_images/1
   # PUT /target_images/1.json
   def update
-    @target_image = TargetImage.find(params[:id])
+    @target_image = ImageManagement::TargetImage.find(params[:id])
 
     respond_to do |format|
       if @target_image.update_attributes(params[:target_image])
@@ -72,7 +72,7 @@ class TargetImagesController < ApplicationController
   # DELETE /target_images/1
   # DELETE /target_images/1.json
   def destroy
-    @target_image = TargetImage.find(params[:id])
+    @target_image = ImageManagement::TargetImage.find(params[:id])
     @target_image.destroy
 
     respond_to do |format|
