--- app/models/base_image.rb
@@ -1,4 +1,0 @@
-class BaseImage < ActiveRecord::Base
-  has_many :image_versions
-  belongs_to :template
-end
