+++ app/models/image_management/base_image.rb
@@ -1,0 +1,6 @@
+module ImageManagement
+ class BaseImage < ActiveRecord::Base
+    has_many :image_versions
+    belongs_to :template
+  end
+end
