+++ app/models/image_management/provider_image.rb
@@ -1,0 +1,5 @@
+module ImageManagement
+  class ProviderImage < ActiveRecord::Base
+    belongs_to :target_image
+  end
+end
