+++ app/models/image_management/target_image.rb
@@ -1,0 +1,6 @@
+module ImageManagement
+  class TargetImage < ActiveRecord::Base
+    belongs_to :image_version
+    has_many :provider_images
+  end
+end
