+++ app/models/image_management/template.rb
@@ -1,0 +1,6 @@
+module ImageManagement
+  class Template < ActiveRecord::Base
+    has_one :base_image
+    has_one :target_image
+  end
+end
