--- app/models/image_version.rb
@@ -1,5 +1,0 @@
-class ImageVersion < ActiveRecord::Base
-  belongs_to :base_image
-  has_many :target_images
-  belongs_to :template
-end
