--- app/models/provider_image.rb
@@ -1,3 +1,0 @@
-class ProviderImage < ActiveRecord::Base
-  belongs_to :target_image
-end
