--- app/models/target_image.rb
@@ -1,4 +1,0 @@
-class TargetImage < ActiveRecord::Base
-  belongs_to :image_version
-  has_many :provider_images
-end
