--- app/models/template.rb
@@ -1,4 +1,0 @@
-class Template < ActiveRecord::Base
-  has_one :base_image
-  has_one :target_image
-end
